+++ test/__tests__/class/format.domain.ts
@@ -1,0 +1,11 @@
+import * as fs from "fs";
+
+import * as Utils from "../../utils";
+
+describe("Format Types", () => {
+  test("format.domain", () => {
+    const generateCode = fs.readFileSync("test/code/class/format.domain/code.ts", { encoding: "utf-8" });
+    const text = Utils.replaceVersionInfo(generateCode);
+    expect(text).toMatchSnapshot();
+  });
+});
