+++ test/__tests__/class/spit-code-test.ts
@@ -1,0 +1,17 @@
+import * as fs from "fs";
+
+import * as Utils from "../../utils";
+
+describe("Split Code", () => {
+  test("types", () => {
+    const generateCode = fs.readFileSync("test/code/class/split/types.ts", { encoding: "utf-8" });
+    const text = Utils.replaceVersionInfo(generateCode);
+    expect(text).toMatchSnapshot();
+  });
+
+  test("apiClient", () => {
+    const generateCode = fs.readFileSync("test/code/class/split/apiClient.ts", { encoding: "utf-8" });
+    const text = Utils.replaceVersionInfo(generateCode);
+    expect(text).toMatchSnapshot();
+  });
+});
