--- test/__tests__/format.domain.ts
@@ -1,16 +1,0 @@
-import * as fs from "fs";
-
-import * as Utils from "../utils";
-
-describe("Format Types", () => {
-  test("format.domain", () => {
-    const generateCode = fs.readFileSync("test/code/class/format.domain/code.ts", { encoding: "utf-8" });
-    const text = Utils.replaceVersionInfo(generateCode);
-    expect(text).toMatchSnapshot();
-  });
-  test("functional-format.domain", () => {
-    const generateCode = fs.readFileSync("test/code/functional/format.domain/code.ts", { encoding: "utf-8" });
-    const text = Utils.replaceVersionInfo(generateCode);
-    expect(text).toMatchSnapshot();
-  });
-});
