+++ test/__tests__/functional/argo-rollout-test.ts
@@ -1,0 +1,11 @@
+import * as fs from "fs";
+
+import * as Utils from "../../utils";
+
+describe("Argo Rollout", () => {
+  test("client.ts", () => {
+    const generateCode = fs.readFileSync("test/code/functional/argo-rollout/client.ts", { encoding: "utf-8" });
+    const text = Utils.replaceVersionInfo(generateCode);
+    expect(text).toMatchSnapshot();
+  });
+});
