+++ test/__tests__/functional/multi-type.test.domain.ts
@@ -1,0 +1,17 @@
+import * as fs from "fs";
+
+import * as Utils from "../../utils";
+
+describe("Multi Type", () => {
+  test("types", () => {
+    const generateCode = fs.readFileSync("test/code/functional/mulit-type-test.domain/types.ts", { encoding: "utf-8" });
+    const text = Utils.replaceVersionInfo(generateCode);
+    expect(text).toMatchSnapshot();
+  });
+
+  test("apiClient", () => {
+    const generateCode = fs.readFileSync("test/code/functional/mulit-type-test.domain/apiClient.ts", { encoding: "utf-8" });
+    const text = Utils.replaceVersionInfo(generateCode);
+    expect(text).toMatchSnapshot();
+  });
+});
