--- test/__tests__/kubernetes-test.ts
@@ -1,16 +1,0 @@
-import * as fs from "fs";
-
-import * as Utils from "../utils";
-
-describe("Kubernetes", () => {
-  test("client-v1.18.5.ts", () => {
-    const generateCode = fs.readFileSync("test/code/class/kubernetes/client-v1.18.5.ts", { encoding: "utf-8" });
-    const text = Utils.replaceVersionInfo(generateCode);
-    expect(text).toMatchSnapshot();
-  });
-  test("functional-client-v1.18.5.ts", () => {
-    const generateCode = fs.readFileSync("test/code/functional/kubernetes/client-v1.18.5.ts", { encoding: "utf-8" });
-    const text = Utils.replaceVersionInfo(generateCode);
-    expect(text).toMatchSnapshot();
-  });
-});
