--- test/__tests__/unknown-schema-domain-test.ts
@@ -1,17 +1,0 @@
-import * as fs from "fs";
-
-import * as Utils from "../utils";
-
-describe("Unknown", () => {
-  test("client.ts", () => {
-    const generateCode = fs.readFileSync("test/code/class/unknown.schema.domain/client.ts", { encoding: "utf-8" });
-    const text = Utils.replaceVersionInfo(generateCode);
-    expect(text).toMatchSnapshot();
-  });
-
-  test("functional-client.ts", () => {
-    const generateCode = fs.readFileSync("test/code/functional/unknown.schema.domain/client.ts", { encoding: "utf-8" });
-    const text = Utils.replaceVersionInfo(generateCode);
-    expect(text).toMatchSnapshot();
-  });
-});
